* src/*: Fix clang warnings.
All of them are minor and belong to one of the following categories:

* signedness casts
* corrected `printf` signatures
* missing `static` keyword for local functions

+++ src/ftgrid.c
@@ -1053,7 +1053,7 @@
 
     /* The floating scale is reversibly adjusted after decomposing it into */
     /* fraction and exponent. Direct bit manipulation is less portable.    */
-    frc = 8 * frexpf( status.scale, &exp );
+    frc = (int)( 8 * frexpf( status.scale, &exp ) );
 
     frc  = ( frc & 3 ) | ( exp << 2 );
     frc += step;
